Added RetrieveMovieDetailsTask which is an AsyncTask to make a call to themoviedb.org. The service call retrieves a list of popular movies and puts JSON response into a string.
The API Key that is required for the call is personal and secure, to use the application
the key has to be added as a resource string.  This can be done in string.xml or another
resource xml file.  Personally I added a file named 'themoviedbAPIKey.xml' with this string.
The string has to be named 'themoviedb_api_key'.

I made a change to the .gitignore so that this file is not checked into GitHub.